(db/snapshot): add Utils.h
