+++ core/src/db/snapshot/Store.h
@@ -13,6 +13,7 @@
 
 #include "db/snapshot/ResourceTypes.h"
 #include "db/snapshot/Resources.h"
+#include "db/snapshot/Utils.h"
 
 #include <stdlib.h>
 #include <time.h>
@@ -35,18 +36,6 @@ namespace milvus {
 namespace engine {
 namespace snapshot {
 
-template <class T, class Tuple>
-struct Index;
-
-template <class T, class... Types>
-struct Index<T, std::tuple<T, Types...>> {
-    static const std::size_t value = 0;
-};
-
-template <class T, class U, class... Types>
-struct Index<T, std::tuple<U, Types...>> {
-    static const std::size_t value = 1 + Index<T, std::tuple<Types...>>::value;
-};
 
 class Store {
  public:
